Moved definitions to source file and only kept declarations in the header
