+++ interfaces/daqp-eigen/CMakeLists.txt
@@ -3,9 +3,9 @@ project(daqp_eigen)
 find_package(Eigen3 3.4.0 REQUIRED NO_MODULE)
 find_package(daqp)
 
-add_executable(00_basic_qp tests/00_basic_qp.cpp)
-add_executable(01_static_func tests/01_static_func.cpp)
-add_executable(02_class tests/02_class.cpp)
+add_executable(00_basic_qp tests/00_basic_qp.cpp daqp.cpp)
+add_executable(01_static_func tests/01_static_func.cpp daqp.cpp)
+add_executable(02_class tests/02_class.cpp daqp.cpp)
 
 set(TARGETS 00_basic_qp 01_static_func 02_class)
